// src/models/DebateTactic.ts
export enum DebateTactic {
  GISH_GALLOP = 'gish_gallop',           // rapid-fire multiple arguments
  STRAWMAN = 'strawman',                  // misrepresent then refute
  AD_HOMINEM = 'ad_hominem',             // attack character/credibility
  APPEAL_TO_EMOTION = 'appeal_to_emotion', // emphasize emotional impact
  APPEAL_TO_AUTHORITY = 'appeal_to_authority', // cite authorities
  FALSE_DILEMMA = 'false_dilemma',       // present only two options
  SLIPPERY_SLOPE = 'slippery_slope',     // chain of consequences
  RED_HERRING = 'red_herring',           // introduce irrelevant topics
  NONE = 'none'                          // avoid fallacious tactics
}